pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chameleon-sre"
version = "0.1.0"
description = "Autonomous Site Reliability Engineer using LangGraph and Ollama"
authors = [
    {name = "MLOps Team", email = "mlops@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
keywords = ["kubernetes", "sre", "langchain", "langgraph", "ollama", "ai-agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Monitoring",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]

dependencies = [
    "langchain>=0.3.0",
    "langchain-ollama>=0.2.0",
    "langgraph>=0.2.0",
    "ollama>=0.4.0",
    "chromadb>=0.5.0",
    "kubernetes>=31.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.10.0",
    "pyyaml>=6.0.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "black>=24.0.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
]

[project.urls]
Homepage = "https://github.com/your-org/chameleon-sre"
Documentation = "https://github.com/your-org/chameleon-sre/docs"
Repository = "https://github.com/your-org/chameleon-sre"
Issues = "https://github.com/your-org/chameleon-sre/issues"

[project.scripts]
chameleon-sre = "src.agent:main"

[tool.setuptools]
packages = ["src", "models", "rag"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=html --cov-report=term-missing"
